update . API 통신 전 테스트 모듈
